added wallpaper augmentation for dataloader

+++ WDNet.py
@@ -298,9 +298,9 @@ class WDNet(object):
                     writer.add_scalar('I_watermark2_Loss', I_watermark2_loss, iter_all)
                     writer.add_scalar('vgg_Loss', vgg_loss, iter_all)
 
-                    watermark_detect = (g_w * g_mask).reshape(-1, 3, 256, 200)
-                    input_image = x_.reshape(-1, 3, 256, 200)
-                    input_mask = mask.reshape(-1, 3, 256, 200)
+                    watermark_detect = (g_w * g_mask).reshape(-1, 3, 256, 200).to(dtype=torch.uint8)
+                    input_image = x_.reshape(-1, 3, 256, 200).to(dtype=torch.uint8)
+                    input_mask = mask.reshape(-1, 3, 256, 200).to(dtype=torch.uint8)
                     img_grid_watermark_detect = torchvision.utils.make_grid(watermark_detect, normalize=True)
                     img_grid_input_img = torchvision.utils.make_grid(input_image, normalize=True)
                     img_grid_input_mask = torchvision.utils.make_grid(input_mask, normalize=True)
